+++ src/net/sourceforge/kolmafia/request/coinmaster/shop/SpinMasterLatheRequest.java
@@ -41,8 +41,8 @@ public abstract class SpinMasterLatheRequest extends CoinMasterShopRequest {
   static {
     Map<Integer, Integer> map = CoinmastersDatabase.getBuyPrices(master);
     for (Entry<Integer, Integer> entry : CoinmastersDatabase.getBuyPrices(master).entrySet()) {
-      int itemId = entry.getKey().intValue();
-      int price = entry.getValue().intValue();
+      int itemId = entry.getKey();
+      int price = entry.getValue();
       AdventureResult cost =
           switch (itemId) {
             default -> FLIMSY_HARDWOOD_SCRAPS.getInstance(price);
